Generate Quiz for Tournament
The Quiz is generated if there are more than 1 signed up students and
if someone calls getRunningTournaments() while it is running. Code had
to be changed to reflect this new change, mostly the fact that now
if it comes time for a Quiz to start but there are no users signed
up for it it is cancelled, which caused some changes in the previous
tests.
Closes #122.